requirements.txt
# .\well-bot_venv\Scripts\Activate.ps1
# python.exe -m pip install --upgrade pip
# ask chatGPT how to permanently add ollama to venv 
# pull and run LLM model from ollama

haystack-ai
ollama-haystack
fastembed-haystack
pyarrow
nltk
qdrant-client

haystack-ai
haystack-ai[integrations-qdrant,integrations-ollama]
qdrant-client==1.15.1
fastapi
uvicorn
pydantic-settings
python-dotenv
pymongo
